Remove Docker container and move to Docker Compose
Instead of running everyhting inside a Docker container, lets use Docker
Compose and deploy different containers for each of the services. This
allows us to get rid of the supervisord, and manage replicas through the
compose file. Also, we switch from NGINX to OpenResty, in order to be
able to dynamically populate the backend servers.