Implement better locking mechanism
When downloading and reading git stuff (TOSCA, catalog) we need to add a
good locking logic, so that two processes do not try to write the same
date at the same time, locking when data is being read, etc. This needs
to be improved, but for now this will work.